refactor: 메인 화면 메뉴 모음 Widget 클래스 분리
